+++ app/javascript/components/admin/site_settings/appearance/Appearance.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import BrandColor from './BrandColor';
+
+export default function Appearance() {
+  return (
+    <BrandColor />
+  );
+}
